requirements.txt
Flask==3.0.2
flask_sqlalchemy==3.1.1
langchain==0.1.11
langchain_community==0.0.27
langchain_openai==0.0.8
python-dotenv==1.0.1
SQLAlchemy==2.0.27
